modify Makefile and add kernel.h to make kernel optional
