--- include/helmholtz.h
@@ -1,8 +1,0 @@
-#ifndef helmholtz_h
-#define helmholtz_h
-
-namespace <{1:exafmm_t}> {
-  void potentialP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value);
-} 
-
-#endif
