+++ include/kernel.h
@@ -1,0 +1,20 @@
+#ifndef kernel_h
+#define kernel_h
+#include "exafmm_t.h"
+
+namespace exafmm_t {
+#if EXAFMM_LAPLACE
+  void potentialP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value);
+  void gradientP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value);
+#ifdef COMPLEX
+  void potentialP2P(RealVec& src_coord, ComplexVec& src_value, RealVec& trg_coord, ComplexVec& trg_value);
+  void gradientP2P(RealVec& src_coord, ComplexVec& src_value, RealVec& trg_coord, ComplexVec& trg_value);
+#endif
+
+#elif EXAFMM_HELMHOLTZ
+  void potentialP2P(RealVec& src_coord, ComplexVec& src_value, RealVec& trg_coord, ComplexVec& trg_value);
+  // void gradientP2P(RealVec& src_coord, ComplexVec& src_value, RealVec& trg_coord, ComplexVec& trg_value);
+#endif
+} 
+
+#endif
