+++ src/laplace.cpp
@@ -47,91 +47,91 @@ namespace exafmm_t {
     return temp;
   }
 
-  void potentialP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value) {
-    simdvec zero((real_t)0);
-    const real_t COEF = 1.0/(2*4*M_PI);   // factor 16 comes from the simd rsqrt function
-    simdvec coef(COEF);
-    int src_cnt = src_coord.size() / 3;
-    int trg_cnt = trg_coord.size() / 3;
-    for(int t=0; t<trg_cnt; t+=NSIMD) {
-      simdvec tx(&trg_coord[3*t+0], 3*(int)sizeof(real_t));
-      simdvec ty(&trg_coord[3*t+1], 3*(int)sizeof(real_t));
-      simdvec tz(&trg_coord[3*t+2], 3*(int)sizeof(real_t));
-      simdvec tv(zero);
-      for(int s=0; s<src_cnt; s++) {
-        simdvec sx(src_coord[3*s+0]);
-        sx = sx - tx;
-        simdvec sy(src_coord[3*s+1]);
-        sy = sy - ty;
-        simdvec sz(src_coord[3*s+2]);
-        sz = sz - tz;
-        simdvec sv(src_value[s]);
-        simdvec r2(zero);
-        r2 += sx * sx;
-        r2 += sy * sy;
-        r2 += sz * sz;
-        simdvec invR = rsqrt(r2);
-        invR &= r2 > zero;
-        tv += invR * sv;
-      }
-      tv *= coef;
-      for(int k=0; k<NSIMD && t+k<trg_cnt; k++) {
-        trg_value[t+k] += tv[k];
-      }
-    }
-    //Profile::Add_FLOP((long long)trg_cnt*(long long)src_cnt*20);
-  }
+  // void potentialP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value) {
+    // simdvec zero((real_t)0);
+    // const real_t COEF = 1.0/(2*4*M_PI);   // factor 16 comes from the simd rsqrt function
+    // simdvec coef(COEF);
+    // int src_cnt = src_coord.size() / 3;
+    // int trg_cnt = trg_coord.size() / 3;
+    // for(int t=0; t<trg_cnt; t+=NSIMD) {
+      // simdvec tx(&trg_coord[3*t+0], 3*(int)sizeof(real_t));
+      // simdvec ty(&trg_coord[3*t+1], 3*(int)sizeof(real_t));
+      // simdvec tz(&trg_coord[3*t+2], 3*(int)sizeof(real_t));
+      // simdvec tv(zero);
+      // for(int s=0; s<src_cnt; s++) {
+        // simdvec sx(src_coord[3*s+0]);
+        // sx = sx - tx;
+        // simdvec sy(src_coord[3*s+1]);
+        // sy = sy - ty;
+        // simdvec sz(src_coord[3*s+2]);
+        // sz = sz - tz;
+        // simdvec sv(src_value[s]);
+        // simdvec r2(zero);
+        // r2 += sx * sx;
+        // r2 += sy * sy;
+        // r2 += sz * sz;
+        // simdvec invR = rsqrt(r2);
+        // invR &= r2 > zero;
+        // tv += invR * sv;
+      // }
+      // tv *= coef;
+      // for(int k=0; k<NSIMD && t+k<trg_cnt; k++) {
+        // trg_value[t+k] += tv[k];
+      // }
+    // }
+    // Profile::Add_FLOP((long long)trg_cnt*(long long)src_cnt*20);
+  // }
 
-  void gradientP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value) {
-    simdvec zero((real_t)0);
-    const real_t COEFP = 1.0/(2*4*M_PI);   // factor 16 comes from the simd rsqrt function
-    const real_t COEFG = -1.0/(4*2*2*6*M_PI);
-    simdvec coefp(COEFP);
-    simdvec coefg(COEFG);
-    int src_cnt = src_coord.size() / 3;
-    int trg_cnt = trg_coord.size() / 3;
-    for(int t=0; t<trg_cnt; t+=NSIMD) {
-      simdvec tx(&trg_coord[3*t+0], 3*(int)sizeof(real_t));
-      simdvec ty(&trg_coord[3*t+1], 3*(int)sizeof(real_t));
-      simdvec tz(&trg_coord[3*t+2], 3*(int)sizeof(real_t));
-      simdvec tv0(zero);
-      simdvec tv1(zero);
-      simdvec tv2(zero);
-      simdvec tv3(zero);
-      for(int s=0; s<src_cnt; s++) {
-        simdvec sx(src_coord[3*s+0]);
-        sx = tx - sx;
-        simdvec sy(src_coord[3*s+1]);
-        sy = ty - sy;
-        simdvec sz(src_coord[3*s+2]);
-        sz = tz - sz;
-        simdvec r2(zero);
-        r2 += sx * sx;
-        r2 += sy * sy;
-        r2 += sz * sz;
-        simdvec invR = rsqrt(r2);
-        invR &= r2 > zero;
-        simdvec invR3 = (invR*invR) * invR;
-        simdvec sv(src_value[s]);
-        tv0 += sv*invR;
-        sv *= invR3;
-        tv1 += sv*sx;
-        tv2 += sv*sy;
-        tv3 += sv*sz;
-      }
-      tv0 *= coefp;
-      tv1 *= coefg;
-      tv2 *= coefg;
-      tv3 *= coefg;
-      for(int k=0; k<NSIMD && t+k<trg_cnt; k++) {
-        trg_value[0+4*(t+k)] += tv0[k];
-        trg_value[1+4*(t+k)] += tv1[k];
-        trg_value[2+4*(t+k)] += tv2[k];
-        trg_value[3+4*(t+k)] += tv3[k];
-      }
-    }
-    //Profile::Add_FLOP((long long)trg_cnt*(long long)src_cnt*27);
-  }
+  // void gradientP2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value) {
+    // simdvec zero((real_t)0);
+    // const real_t COEFP = 1.0/(2*4*M_PI);   // factor 16 comes from the simd rsqrt function
+    // const real_t COEFG = -1.0/(4*2*2*6*M_PI);
+    // simdvec coefp(COEFP);
+    // simdvec coefg(COEFG);
+    // int src_cnt = src_coord.size() / 3;
+    // int trg_cnt = trg_coord.size() / 3;
+    // for(int t=0; t<trg_cnt; t+=NSIMD) {
+      // simdvec tx(&trg_coord[3*t+0], 3*(int)sizeof(real_t));
+      // simdvec ty(&trg_coord[3*t+1], 3*(int)sizeof(real_t));
+      // simdvec tz(&trg_coord[3*t+2], 3*(int)sizeof(real_t));
+      // simdvec tv0(zero);
+      // simdvec tv1(zero);
+      // simdvec tv2(zero);
+      // simdvec tv3(zero);
+      // for(int s=0; s<src_cnt; s++) {
+        // simdvec sx(src_coord[3*s+0]);
+        // sx = tx - sx;
+        // simdvec sy(src_coord[3*s+1]);
+        // sy = ty - sy;
+        // simdvec sz(src_coord[3*s+2]);
+        // sz = tz - sz;
+        // simdvec r2(zero);
+        // r2 += sx * sx;
+        // r2 += sy * sy;
+        // r2 += sz * sz;
+        // simdvec invR = rsqrt(r2);
+        // invR &= r2 > zero;
+        // simdvec invR3 = (invR*invR) * invR;
+        // simdvec sv(src_value[s]);
+        // tv0 += sv*invR;
+        // sv *= invR3;
+        // tv1 += sv*sx;
+        // tv2 += sv*sy;
+        // tv3 += sv*sz;
+      // }
+      // tv0 *= coefp;
+      // tv1 *= coefg;
+      // tv2 *= coefg;
+      // tv3 *= coefg;
+      // for(int k=0; k<NSIMD && t+k<trg_cnt; k++) {
+        // trg_value[0+4*(t+k)] += tv0[k];
+        // trg_value[1+4*(t+k)] += tv1[k];
+        // trg_value[2+4*(t+k)] += tv2[k];
+        // trg_value[3+4*(t+k)] += tv3[k];
+      // }
+    // }
+    // Profile::Add_FLOP((long long)trg_cnt*(long long)src_cnt*27);
+  // }
 
   //! Laplace P2P save pairwise contributions to k_out (not aggregate over each target)
   void kernelMatrix(real_t* r_src, int src_cnt, real_t* r_trg, int trg_cnt, real_t* k_out) {
